Project2/cassandra-writer/requirements.txt
kafka-python==2.0.2
cassandra-driver==3.29.0
